refactor:CategoryConverter 이용해서 Controller에서 Entity 사용 x
