+++ src/main/java/com/blogproject/category/controller/CategoryController.java
@@ -1,20 +1,18 @@
 package com.blogProject.category.controller;
 
-import com.blogProject.category.converter.CategoryConverter;
-import com.blogProject.category.dto.model.CategoryDto;
+import com.blogProject.category.dto.CategoryDto;
 import com.blogProject.category.exception.NameAlreadyExistsException;
 import com.blogProject.category.service.CategoryService;
-import javax.validation.Valid;
 import java.util.List;
-import java.util.stream.Collectors;
+import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -25,14 +23,12 @@ import org.springframework.web.bind.annotation.RestController;
 public class CategoryController {
 
   private final CategoryService categoryService;
-  private final CategoryConverter categoryConverter;
 
   // 카테고리 생성
   @PostMapping
   public ResponseEntity<?> createCategory(@Valid @RequestBody CategoryDto categoryDto) {
     try {
-      CategoryDto responseDto = categoryConverter.entityToDto(
-          categoryService.createCategory(categoryDto.getCategoryName()));
+      CategoryDto responseDto = categoryService.createCategory(categoryDto.getCategoryName());
       return new ResponseEntity<>(responseDto, HttpStatus.CREATED);
     } catch (NameAlreadyExistsException e) {
       return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
@@ -42,18 +38,15 @@ public class CategoryController {
   // 카테고리 전체 보기
   @GetMapping
   public ResponseEntity<?> allCategory() {
-    List<CategoryDto> categoryDtos = categoryService.getAllCategory().stream()
-        .map(categoryConverter::entityToDto)
-        .collect(Collectors.toList());
+    List<CategoryDto> categoryDtos = categoryService.getAllCategory();
     return new ResponseEntity<>(categoryDtos, HttpStatus.OK);
   }
 
   // 카테고리 수정
-  @PatchMapping("/{id}")
+  @PutMapping("/{id}")
   public ResponseEntity<?> updateCategory(@PathVariable Long id,
       @Valid @RequestBody CategoryDto categoryDto) {
-    CategoryDto responseDto = categoryConverter.entityToDto(
-        categoryService.updateCategory(id, categoryDto));
+    CategoryDto responseDto = categoryService.updateCategory(id, categoryDto);
     return new ResponseEntity<>(responseDto, HttpStatus.OK);
   }
 
@@ -65,3 +58,4 @@ public class CategoryController {
   }
 
 }
+
